Процесс моделирования полета луча отделен и реализован в классе TracerLight;
Реалтизовано изменение дискрет вне зоны событий (требуется оптимизация)

+++ src/Application/backend/GIS/map.cpp
@@ -10,7 +10,7 @@ Map::Map()
     lenBlock = 20;
 }
 
-float Map::getMaxH()
+int Map::getMaxH()
 {
     return getCountLayers() * lenBlock;
 }
